variant details for admin

+++ app/Http/Controllers/Backend/VariantDetailsController.php
@@ -2,65 +2,61 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\DataTables\VendorProductVariantDetailsDataTable;
+
+use App\DataTables\VariantDetailsDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
+use App\Models\ProductVariant;
 use App\Models\Variant;
 use App\Models\VariantDetails;
+use App\Models\ProductVariantType;
 use Illuminate\Http\Request;
 
-class VendorProductVariantDetailsController extends Controller
+class VariantDetailsController extends Controller
 {
-    public function index(VendorProductVariantDetailsDataTable $dataTable, $productId, $variantId)
+    public function index(VariantDetailsDataTable $dataTable, $variantId)
     {
-        $product = Product::findOrFail($productId);
-        $variant = Variant::findOrFail($variantId);
-        return $dataTable->render('vendor.product.variant-details.index', compact('product', 'variant'));
+        $variant = ProductVariantType::findOrFail($variantId);
+        return $dataTable->render('admin.product.variant-details.index', compact('variant'));
     }
 
-    public function create($productId, $variantId)
+    public function create($variantId)
     {
-        $variant = Variant::findOrFail($variantId);
-        $product = Product::findOrFail($productId);
-        return view('vendor.product.variant-details.create', compact('variant', 'product'));
+        $variant = ProductVariantType::findOrFail($variantId);
+        return view('admin.product.variant-details.create', compact('variant'));
     }
 
     public function store(Request $request)
     {
         $request->validate([
             'product_variant_type_id' => ['required', 'integer'],
-            'product_id' => ['required', 'integer'],
-            'price' => ['required', 'numeric'],
             'variant_value' => ['required', 'max:200'],
-            'is_default' => ['required'],
             'status' => ['required'],
         ]);
+        $variantValue = new VariantDetails();
+        $variantValue->create($request->all());
 
-        $variantValues = new VariantDetails();
-        $variantValues->create($request->all());
         toastr('Created successfully');
-        return redirect()->route('vendor.product.product-variant-details.index', ['productId' => $request->product_id, 'variantId' => $request->product_variant_type_id]);
+        return redirect()->route('admin.product.variant-details', ['variantId' => $request->product_variant_type_id]);
     }
 
     public function edit($variantDetailsId)
     {
         $variantDetails = VariantDetails::findOrFail($variantDetailsId);
-        return view('vendor.product.variant-details.edit', compact('variantDetails'));
+        return view('admin.product.variant-details.edit', compact('variantDetails'));
     }
 
     public function update(Request $request, $variantDetailsId)
     {
         $request->validate([
-            'price' => ['required', 'numeric'],
             'variant_value' => ['required', 'max:200'],
-            'is_default' => ['required'],
             'status' => ['required'],
         ]);
 
         $variantValue =  VariantDetails::findOrFail($variantDetailsId);
         $variantValue->update($request->all());
         toastr('Updated successfully');
-        return redirect()->route('vendor.product.product-variant-details.index', ['productId' => $variantValue->Variant->product_id, 'variantId' => $variantValue->product_variant_type_id]);
+        return redirect()->route('admin.product.variant-details', ['variantId' => $variantValue->product_variant_type_id]);
     }
     //change status using ajax request--------------------------------------------------
     public function changeStatus(Request $request)
@@ -79,6 +75,10 @@ class VendorProductVariantDetailsController extends Controller
     public function destroy(string $id)
     {
         $variant = VariantDetails::findOrFail($id);
+        $productVarinatDetail = ProductVariant::where('product_variant_detail_id', $id)->first();
+        if ($productVarinatDetail != null) {
+            return response(['status' => 'error', 'message' => 'This variant has product, delete the product first']);
+        }
         $variant->delete();
         return response(['status' => 'success', 'message' => 'Deleted successfully']);
     }
